handle language for second searches after a prefix arg search
- we store lang of a search in leo--results-info
- try to fetch this when clicking or tabbing/enter on any results
- and when we run leo-translate-word again from a results buffer

this enables us to have a whole search session in a given language pair, until
we either quit or set the language again with another prefix argument.

leo-language remains the default language, so if we quit and start again, the
search will use it, as leo--results-info will again be nil.